requirements.txt
discord.py==2.3.1
loguru==0.7.0
selenium==4.10.0
undetected-chromedriver==3.5.0